Refactored backend file names + fixed imports
